+++ frontend/src/views/Cards/AccountsCard.vue
@@ -1,144 +1,146 @@
-<script setup>
-import { defineProps, defineEmits } from "vue";
-
-const props = defineProps({
-  icon: {
-    type: Object,
-    required: false,
-    default: () => ({
-      background: "bg-gradient-success", // 기본값 추가
-    }),
-  },
-  title: {
-    type: String,
-    required: true,
-  },
-  description: {
-    type: String,
-    default: "",
-  },
-  value: {
-    type: [String, Number],
-    default: "",
-  },
-  imgSrc: {
-    type: String,
-    required: true,
-  },
-  img: {
-    type: String,
-    required: false,
-  },
-  isSelected: {
-    type: Boolean,
-    default: false, // 기본값 설정
-  },
-});
-
-const emit = defineEmits(["click"]);
-
-function handleClick() {
-  emit("click");
-}
-</script>
-
 <template>
-  <div
-    class="card position-relative"
-    :class="{ selected: isSelected }"
-    :style="{
-      backgroundImage: `url(${imgSrc})`,
-      backgroundSize: 'cover',
-      backgroundPosition: 'center',
-    }"
-    @click="handleClick"
-  >
-    <!-- 카드 제목 -->
-    <div class="position-relative p-3 text-white">
-      <h6 class="mb-0">{{ title }}</h6>
+    <div v-if="user">
+        <div class="card custom-card" :style="{ backgroundImage: `url(${backgroundImage})` }">
+            <div class="card-body d-flex align-items-end justify-content-end">
+                <div class="d-flex align-items-center">
+                    <div class="icon-container me-2">
+                        <img :src=flagIcon alt="icon" class="icon-image">
+                    </div>
+                    <p class="fs-4 mb-0 me-2">{{ balance }}</p>
+                    <p class="fs-6 mb-0">{{ displayCountry }}</p>
+                </div>
+            </div>
+        </div>
     </div>
-
-    <!-- 카드 안에 표시할 이미지
-    <div v-if="img" class="position-relative p-3 img-container">
-      <img :src="img" alt="Card Image" class="img-fluid rounded" />
-    </div> -->
-
-    <!-- 아이콘과 내용
-    <div class="w-60 mt-2" style="margin-top: 20px">
-      <slot name="icon"></slot>
-    </div> -->
-
-    <!-- 하단 우측 내용 -->
-    <div
-      class="position-absolute bottom-0 end-0 p-3 text-center text-white"
-      style="margin-bottom: -15px; margin-right: 10px"
-    >
-      <hr class="my-1 horizontal dark" />
-      <h6 class="text-s">{{ value }}</h6>
+    <div v-else>
+        <div class="card custom-card" :style="{ bacgroundColor: gray }">
+        </div>
     </div>
-  </div>
 </template>
 
+<script setup>
+import { defineProps, ref, computed, onMounted } from 'vue';
+import myaccountApi from '../../api/myaccountApi';
+import { INTL_LOCALE, CURRENCY_NAMES } from "@/constants/countryCode";
+
+import currencyFormatter from '../../js/currencyFormatter';
+const { formatNumber, formatCurrency } = currencyFormatter;
+
+import { useAuthStore } from '@/stores/auth';
+const auth = useAuthStore();
+const user = computed(() => auth.user);
+
+const isFocused = ref(false);
+const balance = ref(0);
+
+const props = defineProps({
+    // title: {
+    //     type: String,
+    //     default: ''
+    // },
+    // balance: {
+    //     type: Number,
+    //     default: 0
+    // },
+    assetType: {
+        type: String,
+        required: true
+    },
+    // value: {
+    //     type: String,
+    //     default: "",
+    // },
+    // backgroundImage: {
+    //     type: String,
+    //     default: ''
+    // },
+    // icon: {
+    //     type: String,
+    //     default: ''
+    // },
+    // isSelected: {
+    //     type: Boolean,
+    //     default: false
+    // }
+});
+
+//추후에 i18n키값 구분 로직으로 변경
+const backgroundImage = computed(() => `/images/${props.assetType}-money.png`);
+// const icon = computed(() => `/images/${props.currency}.png`);
+const flagIcon = computed(() => {
+    const iconName = props.assetType === 'song-e' ? user.value.countryCode : '0';
+    return `/images/flag_${iconName}.png`;
+});
+const displayCountry = computed(() => {
+    return props.assetType === 'song-e' ? CURRENCY_NAMES[user.value.countryCode] : CURRENCY_NAMES[0];
+});
+
+
+
+const fetchBalance = async () => {
+    if (props.assetType === 'song-e') {
+        myaccountApi.fetchsongeAccountBalance(user.value.songNo).then((fetchedBalance) => {
+            balance.value = formatNumber(fetchedBalance.toFixed(2));
+            // balance.value = formatCurrency(fetchedBalance,INTL_LOCALE[user.value.countryCode],CURRENCY_NAMES[user.value.countryCode]);
+
+        });
+    } else {
+        myaccountApi.fetchkrwAccountBalance(user.value.krwNo).then((fetchedBalance) => {
+            balance.value = formatNumber(fetchedBalance.toFixed(2));
+            // formatCurrency(fetchedBalance,INTL_LOCALE[0],CURRENCY_NAMES[0]);
+        });
+    }
+};
+
+onMounted(() => {
+    fetchBalance();
+});
+
+// fetchBalance 함수를 외부에서 접근 가능하게 만듭니다.
+defineExpose({ fetchBalance });
+
+</script>
+
 <style scoped>
-.card {
-  width: 30vw; /* 뷰포트 너비의 30%로 조절 */
-  max-width: 450px; /* 최대 너비 설정 */
-  aspect-ratio: 3 / 2; /* 비율 유지 */
-  position: relative;
-  border-radius: 20px; /* 카드의 모서리를 둥글게 설정 */
-  overflow: hidden; /* 카드 내용이 카드 영역을 넘지 않도록 설정 */
-  cursor: pointer; /* 카드 클릭 가능하도록 설정 */
-  margin-left: 30px;
-  margin-right: 30px;
+.custom-card {
+    width: 8.56cm;
+    aspect-ratio: 1.585;
+    /* 비율 8.56 / 5.398 */
+    background-size: cover;
 }
 
-.card.selected {
-  border: 3px solid #ffd700;
-  /* 선택된 카드에 녹색 테두리 추가 */
-  box-shadow: 0 0 15px ffd700;
-  /* 선택된 카드에 그림자 효과 추가 */
+@media (max-width: 500px) {
+    .card {
+        width: 90%;
+    }
 }
 
-.img-container {
-  position: absolute;
-  top: 10px; /* 제목 아래에 위치하도록 조절 */
-  left: 10px; /* 카드의 왼쪽 여백 조절 */
-  width: 10vw; /* 뷰포트 너비의 10%로 조절 */
-  max-width: 90px; /* 최대 너비 설정 */
-  height: auto; /* 높이는 비율에 맞춰 조절 */
+.icon-container {
+    /* 뷰포트 너비의 3% */
+    width: 3vw;
+    height: 3vw;
+    /* 최소 크기 설정 */
+    min-width: 20px;
+    min-height: 20px;
+    /* 최대 크기 설정 */
+    max-width: 40px;
+    max-height: 40px;
+    display: flex;
+    align-items: center;
+    justify-content: center;
 }
 
-.img-container img {
-  width: 100%; /* 컨테이너에 맞게 이미지 크기 조절 */
-  height: auto; /* 비율 유지 */
+.icon-image {
+    max-width: 100%;
+    max-height: 100%;
+    object-fit: contain;
 }
 
-/* 텍스트 스타일 */
-h6 {
-  font-size: 2.5vw; /* 뷰포트 너비에 따라 크기 조절 */
-  margin: 0;
-}
-
-/* 하단 우측 내용 글자 크기 조정 */
-.text-s {
-  font-size: 2vw; /* 뷰포트 너비에 따라 크기 조절 */
-}
-/* 미디어 쿼리: 텍스트 크기 조정 */
 @media (max-width: 768px) {
-  h6 {
-    font-size: 1rem; /* 작은 화면에서 글자 크기 감소 */
-  }
-  .text-s {
-    font-size: 0.5rem; /* 작은 화면에서 글자 크기 감소 */
-  }
-}
-
-@media (max-width: 576px) {
-  h6 {
-    font-size: 0.5rem; /* 모바일 화면에서 글자 크기 감소 */
-  }
-  .text-s {
-    font-size: 0.25rem; /* 모바일 화면에서 글자 크기 감소 */
-  }
+    .icon-container {
+        /* 작은 화면에서는 더 큰 비율 적용 */
+        width: 5vw;
+        height: 5vw;
+    }
 }
 </style>
